% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invalidivsacim_est}
\alias{invalidivsacim_est}
\title{This is the main function to compute the estimator in C++}
\usage{
invalidivsacim_est(
  time,
  event,
  stime,
  Z,
  Z_c,
  D_status,
  D_status_c,
  Z_model_mat,
  eps_1,
  D_model_mat,
  eps_2,
  weights
)
}
\arguments{
\item{time}{censored event time}

\item{event}{event indicator}

\item{stime}{unique sorted noncensored event time}

\item{Z}{the instrument}

\item{Z_c}{centered IV}

\item{D_status}{treatment process at each noncensored event time}

\item{D_status_c}{centered treatment process at each noncensored event time}

\item{Z_model_mat}{bread matrix from modeling IV}

\item{eps_1}{influence function from modeling IV}

\item{D_model_mat}{bread matrix from modeling treatment}

\item{eps_2}{influence function from modeling treatment}

\item{weights}{optional weights}
}
\description{
This is the main function to compute the estimator in C++
}
\keyword{internal}
