% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{IV_center}
\alias{IV_center}
\title{This function is to compute the centralized instrument}
\usage{
IV_center(Z, weights)
}
\description{
This function is to compute the centralized instrument
}
\keyword{internal}
