% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{treatment_status}
\alias{treatment_status}
\title{This function is to generate a grid of treatment status}
\usage{
treatment_status(n, K, stime, treatment_init, treatment_shift_time, max.time)
}
\description{
This function is to generate a grid of treatment status
}
\keyword{internal}
