% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{trt_center}
\alias{trt_center}
\title{This function is to compute the centralized treatment process given the covariates Z, L at each sorted unique event time}
\usage{
trt_center(D_status, Z, weights)
}
\description{
This function is to compute the centralized treatment process given the covariates Z, L at each sorted unique event time
}
\keyword{internal}
