% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invalidivsacim_est}
\alias{invalidivsacim_est}
\title{This is the main function to compute the estimator in C++}
\usage{
invalidivsacim_est(
  time,
  event,
  stime,
  Z,
  Z_c,
  D_status,
  D_status_c,
  eps_2,
  Z_c_dot,
  weights
)
}
\description{
This is the main function to compute the estimator in C++
}
\keyword{internal}
