% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{IV_center}
\alias{IV_center}
\title{This function is to compute the centralized instrument given the covariates L}
\usage{
IV_center(Z, L = NULL)
}
\description{
This function is to compute the centralized instrument given the covariates L
}
\keyword{internal}
