% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ivsacim_est}
\alias{ivsacim_est}
\title{This is the main function to compute the estimator in C++}
\usage{
ivsacim_est(time, event, stime, Zc, D_status, eps_2, Zc_dot)
}
\description{
This is the main function to compute the estimator in C++
}
\keyword{internal}
